+++ Components/UnitFrames/Units/Party.lua
@@ -1108,6 +1108,7 @@ local style = function(self, unit)
 	auras.sortMethod = db.AurasSortMethod
 	auras.sortDirection = db.AurasSortDirection
 	auras.reanchorIfVisibleChanged = true
+	auras.allowCombatUpdates = true
 	auras.CreateButton = ns.AuraStyles.CreateButton
 	auras.PostUpdateButton = ns.AuraStyles.PartyPostUpdateButton
 	auras.CustomFilter = ns.AuraFilters.PartyAuraFilter -- classic